Use float64 by deafult, more implementation for particle.go
* A lot more is going on in this commit than I would have liked.
* First, ParticleSystem & ParticleManager have most of their
  imprementation in place now.
* Next, and this is large (not only for the commit, but for the API in
  general) is that float64 is passed around wherever possible instead of
  float32. This cleans up casting that needs to take place when passing
  to math functions, for instance. It also allows := 0.0 to be used
  without having to cast it as was previous needed (:= float32(0.0)).
  Overall, I think this is for the best. I /think/ float32 is still
  expected though for the structs that are passed directly to the C
  functions (such as Quad, Vertex, and the like.
* Therefore, all unneeded casts are removed as well.
* Also added, is the ability to pass float64 or float32 arguments when
  the argument is optional and passed by a vararg interface{}.
* Finally, the tutorials are updated to use float64 as well as using
  defer to cleanup main.

+++ color.go
@@ -4,7 +4,7 @@ import (
 	"math"
 )
 
-func colorClamp(x *float32) {
+func colorClamp(x *float64) {
 	if *x < 0.0 {
 		*x = 0.0
 	}
@@ -14,7 +14,7 @@ func colorClamp(x *float32) {
 }
 
 type ColorRGB struct {
-	R, G, B, A float32
+	R, G, B, A float64
 }
 
 type Color interface {
@@ -23,7 +23,7 @@ type Color interface {
 	GetHWColor() Dword
 }
 
-func NewColorRGB(r, g, b, a float32) ColorRGB {
+func NewColorRGB(r, g, b, a float64) ColorRGB {
 	var c ColorRGB
 
 	c.R = r
@@ -80,15 +80,15 @@ func (c ColorRGB) NotEqual(c2 ColorRGB) bool {
 	return c.R != c2.R && c.G != c2.G && c.B != c2.B && c.A != c2.A
 }
 
-func (c ColorRGB) DivScalar(scalar float32) ColorRGB {
+func (c ColorRGB) DivScalar(scalar float64) ColorRGB {
 	return NewColorRGB(c.R/scalar, c.G/scalar, c.B/scalar, c.A/scalar)
 }
 
-func (c ColorRGB) MulScalar(scalar float32) ColorRGB {
+func (c ColorRGB) MulScalar(scalar float64) ColorRGB {
 	return NewColorRGB(c.R*scalar, c.G*scalar, c.B*scalar, c.A*scalar)
 }
 
-func (c *ColorRGB) MulScalarEqual(scalar float32) *ColorRGB {
+func (c *ColorRGB) MulScalarEqual(scalar float64) *ColorRGB {
 	c.R *= scalar
 	c.G *= scalar
 	c.B *= scalar
@@ -105,10 +105,10 @@ func (c *ColorRGB) Clamp() {
 }
 
 func (c *ColorRGB) SetHWColor(col Dword) {
-	c.A = float32(col>>24) / 255.0
-	c.R = float32((col>>16)&0xFF) / 255.0
-	c.G = float32((col>>8)&0xFF) / 255.0
-	c.B = float32(col&0xFF) / 255.0
+	c.A = float64(col>>24) / 255.0
+	c.R = float64((col>>16)&0xFF) / 255.0
+	c.G = float64((col>>8)&0xFF) / 255.0
+	c.B = float64(col&0xFF) / 255.0
 }
 
 func (c ColorRGB) GetHWColor() Dword {
@@ -116,10 +116,10 @@ func (c ColorRGB) GetHWColor() Dword {
 }
 
 type ColorHSV struct {
-	H, S, V, A float32
+	H, S, V, A float64
 }
 
-func NewColorHSV(h, s, v, a float32) ColorHSV {
+func NewColorHSV(h, s, v, a float64) ColorHSV {
 	var c ColorHSV
 
 	c.H = h
@@ -176,15 +176,15 @@ func (c ColorHSV) NotEqual(c2 ColorHSV) bool {
 	return c.H != c2.H && c.S != c2.S && c.V != c2.V && c.A != c2.A
 }
 
-func (c ColorHSV) DivScalar(scalar float32) ColorHSV {
+func (c ColorHSV) DivScalar(scalar float64) ColorHSV {
 	return NewColorHSV(c.H/scalar, c.S/scalar, c.V/scalar, c.A/scalar)
 }
 
-func (c ColorHSV) MulScalar(scalar float32) ColorHSV {
+func (c ColorHSV) MulScalar(scalar float64) ColorHSV {
 	return NewColorHSV(c.H*scalar, c.S*scalar, c.V*scalar, c.A*scalar)
 }
 
-func (c *ColorHSV) MulScalarEqual(scalar float32) *ColorHSV {
+func (c *ColorHSV) MulScalarEqual(scalar float64) *ColorHSV {
 	c.H *= scalar
 	c.S *= scalar
 	c.V *= scalar
@@ -201,13 +201,13 @@ func (c *ColorHSV) Clamp() {
 }
 
 func (c *ColorHSV) SetHWColor(col Dword) {
-	c.A = float32(col>>24) / 255.0
-	r := float32((col>>16)&0xFF) / 255.0
-	g := float32((col>>8)&0xFF) / 255.0
-	b := float32(col&0xFF) / 255.0
+	c.A = float64(col>>24) / 255.0
+	r := float64((col>>16)&0xFF) / 255.0
+	g := float64((col>>8)&0xFF) / 255.0
+	b := float64(col&0xFF) / 255.0
 
-	minv := float32(math.Min(math.Min(float64(r), float64(g)), float64(b)))
-	maxv := float32(math.Max(math.Max(float64(r), float64(g)), float64(b)))
+	minv := math.Min(math.Min(r, g), b)
+	maxv := math.Max(math.Max(r, g), b)
 	delta := maxv - minv
 
 	c.V = maxv
@@ -239,7 +239,7 @@ func (c *ColorHSV) SetHWColor(col Dword) {
 }
 
 func (c ColorHSV) GetHWColor() Dword {
-	var r, g, b float32
+	var r, g, b float64
 	if c.S == 0 {
 		r = c.V
 		g = c.V
@@ -249,7 +249,7 @@ func (c ColorHSV) GetHWColor() Dword {
 		if xh == 6 {
 			xh = 0
 		}
-		i := float32(math.Floor(float64(xh)))
+		i := math.Floor(xh)
 		p1 := c.V * (1 - c.S)
 		p2 := c.V * (1 - c.S*(xh-i))
 		p3 := c.V * (1 - c.S*(1-(xh-i)))
